requirements-dev.txt
-r requirements.txt

pytest-django==4.7.0
pytest==7.4.3
black==23.12.0
flake8==6.1.0
pydantic==2.5.0
djangorestframework==3.14.0
pandas
xlwt